Calculate end of new sequences automatically
When the user has recorded the first sequence, we know how long this
sequence is. The user shouldn't have to press the vocal harmony button
again, when he wants to stop a subsequent recording.
That's why we calculate the time when the sequence should stop
automatically and schedule the stop event in the future.