Improve scheduler and outsource 'heavy' computation
Calculating all the delta times shouldn't be done when the musician
expects the tones to sound. Until now, you could hear a small delay when
having recorded the first sequence and having pressed the vocal harmony
button, so that the looper could start. This is because just right at this
moment all the delta times were calculated.
Instead, we now calculate those times whenever we add a new message
to the array, e. g. whenever the keyboarder presses a key.