Change condition when sequence is registered
Only accept a sequence and move a channel up if the array of messages
contains any Note on or Note off message. Otherwise, all of the recorded
messages in the last sequence will be deleted and new messages can be
collected. This behaviour prevents that switching between instruments is
registered as a sequence (but without having pressed any key on the
keyboard).