Add handling for "cut-off" notes
The keyboard player may press a key but not release it before the
sequence ends. This leads to unwanted notes being played constantly in
the background since there are some NOTE_OFF messages missing in the
recorded channel.
That's why we now store the keys that are missing a NOTE_OFF message.
If, later on, we encounter a NOTE_OFF message for a note registered in
the 'missing notes' array we know that we should put it in the according
channel and not in the current one.