Change timing strategy
While playing, the MIDI looper got slightly out of sync over a longer
period of time (more than 2 minutes). This is because we are
calculating the sequence duration only once based on the bpm value.
It is way more accurate to rely on the MIDI Clock messages (0xF8) the
keyboards sends 24 times per quarter note. Thus, the playback starts
every n ticks. This way, we can be sure that the playback is in sync, at
least every time the playback starts again (every x measures).

(!) Note that this change will temporarily break the calculation for the
stuck notes we previously introduced. While playing, you will probably
hear stuck notes again with this commit.