Improve MIDI time handling and handle stuck notes
This fixes the stuck notes again (this functionality did not work
anymore after having changed the timing strategy in a recent commit).
Furthermore, the MIDI messages timing was changed for easier use
in the code: instead of saving the delta times between two MIDI messages,
we only store the time since the sequence started in every message.
This makes it way easier to handle the stuck notes when we have to put
a MIDI message into another sequence.
Also refactored the code, especially when it comes to nested if-clauses
and big functions that were separated into more meaningful entities.