Improve stuck notes handling
We now handle every note as soon as it drops in, instead of having to
iterate over all of the messages in one sequence when we actually want
to start the playback as soon as possible.
The logic for handling the stuck notes is now bundled in one function.
We now also make out the case when the keyboard player presses a key and
does not release it in the next sequence (organ point).